taggen_base_image/requirements.txt
pyyaml
flask
gunicorn
watchdog 
numpy 
scipy 
scikit-learn 
tqdm 
envoy
